infrastructure/requirements.txt
attrs==23.1.0
aws-cdk-lib==2.96.2
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
bandit==1.7.5
boto3==1.28.50
botocore==1.31.50
cattrs==23.1.2
cdk-nag==2.27.144
constructs==10.2.70
exceptiongroup==1.1.3
gitdb==4.0.10
GitPython==3.1.37
importlib-resources==6.0.1
iniconfig==2.0.0
jmespath==1.0.1
jsii==1.89.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==23.1
pbr==5.11.1
pluggy==1.3.0
publication==0.0.3
Pygments==2.16.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
rich==13.5.3
s3transfer==0.6.2
six==1.16.0
smmap==5.0.1
stevedore==5.1.0
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.7.1
urllib3==1.26.16
zipp==3.17.0
